Removes branch check from script and adds to workflow
The branch check in the `update_dockerhub_description.py` script is redundant since the Github workflow can handle this. Removes the branch check from the Python script and adds a conditional statement to the Github workflow to only run on the main branch. This simplifies the python script and centralizes the branch check in the workflow.