Updates Dockerfile to reference correct path (#274)
* Adds base Dockerfile and CI for image deployment

Sets up a base Dockerfile for development containers with essential tools like Python, Node.js, and Docker CLI.

Also introduces a GitHub Actions workflow to automatically build and push the Docker image to the GitHub Container Registry every two months, using a date-based tag.

* hotfix

* Removes dev requirements installation

Removes the installation of development requirements from the base Dockerfile.

This change likely reflects a shift in the build process,
potentially to optimize image size or to handle dev dependencies
in a separate, more targeted manner.

* Removes dev requirements installation

Removes the installation of development requirements from the base Dockerfile.

This change likely reflects a shift in the build process,
potentially to optimize image size or to handle dev dependencies
in a separate, more targeted manner.