Updates Dockerfile to reference correct path (#275)
* Adds base Dockerfile and CI for image deployment

Sets up a base Dockerfile for development containers with essential tools like Python, Node.js, and Docker CLI.

Also introduces a GitHub Actions workflow to automatically build and push the Docker image to the GitHub Container Registry every two months, using a date-based tag.

* hotfix

* Removes dev requirements installation

Removes the installation of development requirements from the base Dockerfile.

This change likely reflects a shift in the build process,
potentially to optimize image size or to handle dev dependencies
in a separate, more targeted manner.

* Removes dev requirements installation

Removes the installation of development requirements from the base Dockerfile.

This change likely reflects a shift in the build process,
potentially to optimize image size or to handle dev dependencies
in a separate, more targeted manner.

* Updates dev container image repository

Changes the Docker image repository to use the repository owner
instead of the full repository name.

This ensures the image is pushed to the correct location within
the GitHub Container Registry.