Enhances UI with theme toggle and support links
Improves the user interface by adding a theme toggle for light/dark mode.

Also introduces prominent links to support the project via GitHub and PayPal,
enhancing discoverability and user engagement.

Updates the page footer to emphasize the open-source nature of the project
and streamlines the feedback mechanism.

The tooltip component is updated to use a portal.